requirements.txt
flask
sentence-transformers
faiss-cpu
pypdf
python-dotenv
google-genai
numpy
reportlab